requirements.txt
torch
torchvision
scikit_image
streamlit
numpy
opencv_python_headless
Pillow

